+++ R/theme.r
@@ -1,6 +1,7 @@
 #' @title Function that fits the theme model
 #' @description Calculates the components to predict all the dependent variables.
 #' @export
+#' @importFrom ade4 dudi.mix dudi.pca
 #' @param formula an object of class "\code{\link[=multivariateFormula]{MultivariateFormula}}" (or one that can be coerced
 #' to that class): a symbolic description of the model to be fitted. The details
 #' of model specification are given under Details.